refactor: Move probe utilities to common location
Move pod_is_ready() and get_restart_counts() from vLLM-specific
location to tests/model_serving/model_runtime/utils.py for reuse
across all model serving runtimes.

Changes:
- Add get_restart_counts() and pod_is_ready() to common utils
- Remove these functions from vllm/probes/utils.py
- Update vLLM probe tests to import from common utils
- Update MLServer probe tests to import from common utils
- Made docstrings generic (removed "vLLM" references)

This enables sharing these utilities across MLServer, vLLM, Triton,
and future runtimes without duplication.
