+++ tests/model_serving/model_runtime/vllm/probes/test_vllm_probes.py
@@ -8,11 +8,10 @@ from tests.model_serving.model_runtime.vllm.cpu.cpu_x86.constant import (
     CPU_X86_SERVING_ARGUMENT,
     OPT_125M_MODEL_PATH,
 )
+from tests.model_serving.model_runtime.utils import get_restart_counts, pod_is_ready
 from tests.model_serving.model_runtime.vllm.probes.utils import (
     exec_vllm_health_check,
     get_probe,
-    get_restart_counts,
-    pod_is_ready,
     resolve_http_get,
 )
 from utilities.constants import KServeDeploymentType
