+++ tests/model_serving/model_runtime/vllm/probes/utils.py
@@ -135,27 +135,3 @@ def exec_vllm_health_check(pod: Pod, http_get: dict[str, Any]) -> str:
         if last_status == "200":
             return last_status
     return last_status
-
-
-def get_restart_counts(pod: Pod) -> dict[str, int]:
-    """Return container restart counts for the pod.
-
-    Args:
-        pod: Predictor pod for the vLLM InferenceService.
-
-    Returns:
-        Mapping of container name to restartCount.
-    """
-    return {container.name: container.restartCount for container in (pod.instance.status.containerStatuses or [])}
-
-
-def pod_is_ready(pod: Pod) -> bool:
-    """Return True when the pod Ready condition is True.
-
-    Args:
-        pod: Predictor pod for the vLLM InferenceService.
-    """
-    for condition in pod.instance.status.conditions or []:
-        if condition.type == "Ready":
-            return condition.status == "True"
-    return False
